# new-server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BKRM Retail Chain Management",
        "description": "L5 Swagger API for BKRM Retail Chain Management",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/shift": {
            "get": {
                "tags": [
                    "Shift"
                ],
                "summary": "Get all shifts",
                "operationId": "getAllShifts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Shift"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Shift"
                ],
                "summary": "Create a new shift",
                "operationId": "createShift",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateShiftInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shift"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shift/{shift_id}": {
            "get": {
                "tags": [
                    "Shift"
                ],
                "summary": "Get a shift",
                "operationId": "getShift",
                "parameters": [
                    {
                        "name": "shift_id",
                        "in": "path",
                        "description": "Shift ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shift"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Shift"
                ],
                "summary": "Update a shift",
                "operationId": "updateShift",
                "parameters": [
                    {
                        "name": "shift_id",
                        "in": "path",
                        "description": "Shift ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpsertShiftInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shift"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shift"
                ],
                "summary": "Delete a shift",
                "operationId": "deleteShift",
                "parameters": [
                    {
                        "name": "shift_id",
                        "in": "path",
                        "description": "Shift ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shift"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpsertTime": {
                "required": [
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Shift": {
                "required": [
                    "id",
                    "name",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "branch_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "start_time": {
                        "type": "string"
                    },
                    "end_time": {
                        "type": "string"
                    }
                },
                "type": "object",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/UpsertTime"
                    }
                ]
            },
            "UpsertShiftInput": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "start_time": {
                        "type": "string"
                    },
                    "end_time": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateShiftInput": {
                "required": [
                    "name",
                    "start_time",
                    "end_time"
                ],
                "type": "object",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/UpsertShiftInput"
                    }
                ]
            },
...
        }
    },
...
}